[IMP] web_chatter_position: split chatter_responsive to let the user decide
When using side mode, the chatter is by default responsive by addind the isChatterAside and o-aside class. This can cause some 'inconveniencies' if the chatter contains a long text, making the form view almost unreadable while helping reading the chatter. So this PR simply removes the responsive when used in side mode.